pipelines/longterm/requirements.txt
pylint
pytest
boto3
pandas
python-dotenv
pyarrow